Switch clnors data from float[2] to short[2].
Suggested by Campbell, saves memory and does not make things slower (perhaps even slightly faster),
even though it asks for a few more operations...

Note the way alpha factor is handled is a bit different now, since with shorts it's simpler to
always keep factor-like value in the 'virtual' [-1, 1] (real [-SHRT_MAX, SHRT_MAX]) range.

Code is a bit more verbose, since we do not have much math helpers for short 2d vectors,
but this is a detail that can be solved later.