First use of txdata to transfer custom normals...
Had to add a pre/post process helper to txdata, since with clnors
we want to operate on regular normals data, but store results in
'compressed' clnor data.

Note things seems to be working basically, but the whole branch needs to
re-check how/when it's using (custom)lnors (same goes for master, actually!).