More cleanup, simplied set_clnor funcs, reworked mix feature of setsplinor modifier
Now, `BKE_mesh_normals_loop_custom_set()` & co do not take a factor value anymore.
It could make perfomances slightly better, but it was adding some complexity and fuzzyness
in an area that do not need that. So now, caller code is expected to handle
mixing itself, if needed.

And SetSplitNormal modifier now has a real mixing handling, quite similar to
the one of txdata.